backend/requirements.txt
fastapi == 0.118.2
uvicorn[standard]
pydantic == 2.12.0
python-multipart
beautifulsoup4 == 4.14.2
PyPDF2 == 3.0.1

langchain==0.3.27
langchain-community==0.3.31
langchain-openai == 0.3.35
chromadb == 1.1.1
pandas == 2.3.3
streamlit == 1.50.0
python-dotenv==1.0.1

